+++ CHANGELOG.md
@@ -13,3 +13,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### 0.0.7
 - change README.d && CHANGELOG.md
 - change zhe image for preview
+
+### 0.0.9
+- keep the tag's sequence
+- add the config of "indent_root" for html
+- add the config of "force_format" for html, make sure the "template" tag can be format normally
+- add the config of "html_indent_root"